services/git-service/requirements.txt
# AutoGraph v3 - Git Service Dependencies

# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Validation
pydantic==2.10.0
pydantic-settings==2.6.0

# Git
pygit2==1.16.0

# HTTP Client
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
